main page merge w ppops & adapt
