--- src/client/Main/components/SectionBanner/SectionBanner.scss
@@ -1,28 +1,0 @@
-@import 'src/shared/styles/scss/shared';
-
-.section-banner{
-  &--container {
-    background: $background-color--secondary;
-    border-radius: $border-radius;
-  }
-
-  &--description {
-    height: 120px;
-    @include flex(space-between,center);
-    margin: 0 40px;
-  }
-
-  &--text {
-    color: $text-color--main;
-    font-weight: bold;
-    font-family: $font-main;
-    font-size: 24px;
-    margin: 0;
-  }
-  &--button{
-    width: 170px;
-    height: 47px;
-    text-transform: uppercase;
-    font-size: 18px;
-  }
-}
