--- src/client/Main/components/SectionTourList/SectionTourList.scss
@@ -1,5 +1,0 @@
-// export const CarouselCurrentTournamentsTitle = {
-//   color: 'white',
-//   fontWeight: 'bold',
-//   margin: '40px 0',
-// };
