--- src/client/Main/components/SectionTourList/TourCard/TourCard.scss
@@ -1,75 +1,0 @@
-@import 'src/shared/styles/scss/shared';
-
-.flex-row {
-    @include flex(space-between, center);
-    flex-direction: row
-}
-
-.tour-card {
-    background-color: $primary-color;
-    border-radius: $border-radius;
-    margin: 10px;
-
-    &--banner {
-        height: 170px;
-        border-radius: $border-radius;
-        background-repeat: no-repeat;
-        background-size: cover;
-    }
-
-    &--description {
-        padding: 20px;
-        @include flex(space-between);
-        flex-direction: column;
-        height: 150px;
-    }
-
-    &--name_container {
-        @include flex(space-between, flex-start)
-    }
-
-    &--name {
-        @include font-source-sans(21px, white);
-        width: 50%;
-        word-wrap: break-word;
-        overflow: hidden;
-        margin: 0;
-        text-align: flex-start;
-    }
-
-    &--name_prize {
-        @include font-source-sans(false, #747BAA);
-        width: 75px;
-    }
-
-    &--sum {
-        @include font-source-sans(18x, #8B79FF, bold);
-        @include flex(center);
-        margin: 0
-    }
-
-    &--go {
-        @include font-source-sans(19px, #99E47F);
-        color: #99E47F;
-        font-size: 19px;
-        align-items: center;
-        margin: 0
-    }
-
-    &--slots {
-        @include font-source-sans(false, #747BAA);
-        width: 75px;
-    }
-
-    &--button {
-        width: 175px;
-        height: 47px;
-        text-transform: uppercase;
-        font-size: 18px;
-    }
-
-    &--checkIn {
-        @include font-source-sans(false, #747BAA);
-        width: 75px;
-    }
-}
