--- src/client/Main/components/SliderSection/components/SliderContainer/SliderContainer.scss
@@ -1,16 +1,0 @@
-@import 'src/shared/styles/scss/shared';
-
-.slider{
-    &-container{
-        width: 1260px;
-        height: 390px;
-        overflow: hidden;
-   }
-
-    &-items{
-        @include flex(space-between,center);
-        transition: all 500ms ease 0s;
-        transition-timing-function:cubic-bezier(45,-0.67,.53,1.65)
-
-    }
-};
