--- src/client/Main/components/SliderSection/components/SliderItem/SliderItem.scss
@@ -1,30 +1,0 @@
-@import "src/shared/styles/scss/shared";
-
-.slider-item{
-    height: 450px;
-    width: 1260px;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    justify-content: center;
-    background-repeat: no-repeat;
-    background-size: cover;
-    box-shadow: inset 0 3px 35px 33px #26254F;
-
-    &--header{
-        @include font-source-sans(50px, #fff, bold);
-        text-transform: uppercase;
-        margin: 15px;
-      }
-
-    &--description {
-        @include font-source-sans(32px, #fff, bold);
-        margin: 15px;
-    }
-
-      &--button{
-        width: 205px;
-        height: 60px;
-        margin-left: 15px;
-      }
-}
