--- src/client/Main/pages/MainPage/MainPage.scss
@@ -1,12 +1,0 @@
-@import 'src/shared/styles/scss/shared.scss';
-
-.main-page {
-    text-align: -webkit-center;
-    background-color: #26254F;
-    width: 100%;
-    padding-bottom: 70px;
-
-    &--container {
-        width: 1260px;
-    }
-  };
