+++ src/client/mainPage/components/SectionBanner/SectionBanner.module.scss
@@ -1,0 +1,31 @@
+@import 'src/shared/styles/scss/shared';
+
+.sectionBannerContainer {
+    background: $background-color--secondary;
+    border-radius: $border-radius;
+  }
+
+.sectionBannerDescription {
+    height: 120px;
+    @include flex(space-between,center);
+    margin: 0 40px;
+    @media screen and (max-width: 650px) {
+      justify-content: center;
+    }
+  }
+
+.sectionBannerText {
+    text-align: center;
+    color: $text-color--main;
+    font-weight: bold;
+    font-family: $font-main;
+    font-size: 24px;
+    margin: 0;
+  }
+  
+.sectionBannerButton{
+    width: 170px;
+    height: 47px;
+    text-transform: uppercase;
+    font-size: 18px;
+  }
