+++ src/client/mainPage/components/SectionBanner/components/Banner/Banner.module.scss
