+++ src/client/mainPage/components/SectionBanner/components/Banner/index.js
