+++ src/client/mainPage/components/SectionBanner/components/InfoList/InfoList.module.scss
@@ -1,0 +1,28 @@
+@import 'src/shared/styles/scss/shared';
+
+.infoList{
+    @include flex(space-evenly, center);
+    width: 100%;
+}
+.infoListContainer{
+    @include flex(space-evenly, center);
+    width: 270px;
+    height: 170px;
+    background: rgba(54, 72, 134, 0.6);
+    border-radius: $border-radius;
+    padding: 0 4%;
+    }
+.infoListDescription{
+    margin: 0;
+    text-align: center;
+    @include font-source-sans(false, #8988B6);
+    }
+    
+.infoListText {
+    color: $text-color--main;
+    font-weight: bold;
+    font-family: $font-main;
+    font-size: 24px;
+    margin: 0;
+    }
+
