+++ src/client/mainPage/components/SectionBanner/components/InfoList/index.js
