+++ src/client/mainPage/components/SectionBanner/index.js
