+++ src/client/mainPage/components/SectionTourList/SectionTourList.module.scss
@@ -1,0 +1,11 @@
+@import 'src/shared/styles/scss/shared.scss';
+
+.tourList{
+    @include flex(space-between, flex-start);
+
+    @media screen and (max-width: 650px) {
+    @include flex(center, flex-start)
+    }
+}
+
+
