+++ src/client/mainPage/components/SectionTourList/TourCard/TourCard.module.scss
@@ -1,0 +1,98 @@
+@import 'src/shared/styles/scss/shared';
+
+.flex_row {
+    @include flex(space-between);
+    flex-direction: row;
+    height: 40px;
+}
+
+.tour_card {
+    background-color: $primary-color;
+    border-radius: $border-radius;
+    width: 300px;
+    margin-bottom: 20px;
+ 
+
+    &__banner {
+        height: 145px;
+        border-radius: $border-radius;
+        background-repeat: no-repeat;
+        background-size: cover;
+    }
+
+    &__description {
+        padding: 10px 20px;
+        @include flex(space-evenly);
+        flex-direction: column;
+        height: 185px;
+    }
+
+    &__text{
+        @include font-source-sans(12px, #747BAA, 400);
+     }
+
+    &__name {
+        @include font-source-sans(18px, white, 700);
+        width: 50%;
+        word-wrap: break-word;
+        overflow: hidden;
+        margin: 0;
+        text-align: flex-start;
+        line-height: 1.19;
+        letter-spacing: normal
+    }
+
+    &__info {
+        @include flex(space-around,center);
+        flex-direction: column;
+        @include font-source-sans(false, #747BAA);
+        width: 60px;
+        
+    }
+    &__priseCurrency {
+        @include flex(center);
+    }
+
+    &__priseSum {
+        @include font-source-sans(18px, #8B79FF, 700);
+        @include flex(center);
+        margin-left: 3px
+    }
+
+    &__go {
+        @include font-source-sans(15px, #99E47F, 700);
+        color: #99E47F;
+        align-items: center;
+        margin: 0;
+        align-self: center;
+
+    }
+    
+    &__slots {
+        @include font-source-sans(15px, #ffffff,bold);
+    }
+
+    &__sum {
+        @include font-source-sans(15px, #8B79FF, 700);
+        @include flex(center);
+        margin-left: 3px
+    }
+
+    &__button {
+        width: 175px;
+        height: 47px;
+        text-transform: uppercase;
+        font-size: 18px;
+    }
+    &__buttonSection{
+        @include flex(space-between);
+        flex-direction: row;
+        height: 40px;
+        padding: 17px 0;
+    }
+
+    &__checkIn {
+        @include font-source-sans(false, #747BAA);
+        width: 75px;
+    }
+}
