+++ src/client/mainPage/components/SectionTourList/TourCard/index.js
