+++ src/client/mainPage/components/SliderSection/components/SliderContainer/SliderContainer.module.scss
@@ -1,0 +1,14 @@
+@import 'src/shared/styles/scss/shared';
+
+.sliderContainer{
+        width: 100%;
+        height: 390px;
+        overflow: hidden;
+   }
+
+.sliderItems{
+    @include flex(space-between,center);
+    transition: all 500ms ease 0s;
+    transition-timing-function:cubic-bezier(45,-0.67,.53,1.65)
+
+}
