+++ src/client/mainPage/components/SliderSection/components/SliderContainer/index.js
