+++ src/client/mainPage/components/SliderSection/components/SliderDots/SliderDots.scss
