+++ src/client/mainPage/components/SliderSection/components/SliderDots/index.js
