+++ src/client/mainPage/components/SliderSection/components/SliderItem/SliderItem.jsx
