+++ src/client/mainPage/components/SliderSection/components/SliderItem/SliderItem.scss
@@ -1,0 +1,24 @@
+@import "src/shared/styles/scss/shared";
+
+.slider-item{
+    height: 450px;
+    width: 1260px;
+    background-repeat: no-repeat;
+    box-shadow: inset 0 3px 35px 33px #26254F;
+    padding-top:90px;
+
+    &--header{
+        @include font-source-sans(50px, #fff, bold);
+      }
+
+    &--description {
+        @include font-source-sans(28px, #fff, 700);
+        padding: 8px 0 54px;
+
+    }
+
+      &--button{
+        width: 198px;
+        height: 60px;
+      }
+}
