+++ src/client/mainPage/components/SliderSection/components/SliderItem/index.js
