+++ src/client/mainPage/components/SliderSection/index.js
