+++ src/client/mainPage/pages/MainPage/MainPage.module.scss
@@ -1,0 +1,5 @@
+@import 'src/shared/styles/scss/style.scss';
+
+.mainPage {
+    padding-bottom: 70px;
+  };
