+++ src/client/mainPage/pages/MainPage/index.js
