+++ src/client/mainPage/pages/MainPage/pageProps;.js
@@ -1,0 +1,362 @@
+import bannerEnticingBg from '../../../../assets/images/banners/bg_banner1.png';
+import bannerAdvantageBg from '../../../../assets/images/banners/bg_banner2.png';
+import { v4 as uuidv4 } from 'uuid';
+
+export const pageProps = {
+  sectionBannerEnticing: {
+    bg: bannerEnticingBg,
+    title: 'Why team better then solo play?',
+    text: 'Create your own team or join to your friends team',
+  },
+  sectionBannerAdvantage: {
+    bg: bannerAdvantageBg,
+    title: 'Join useTeam Right now',
+    text: 'More than 1000 tournaments every month. Check your skills!',
+  },
+};
+
+export const TournamentsData = [
+  {
+    name: 'useTeam',
+    id: uuidv4(),
+    url_for_banner:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606327391/useTeam/tournament1_r2lflx.png',
+    url_for_card:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606327391/useTeam/tournament1_r2lflx.png',
+    description: 'Its your way to become Dota 2 pro-player',
+    status: 'announce',
+    mode: '5x5',
+    entry_from: '20',
+    entry_pool: 'Real money',
+    prise_pool: 'Real money',
+    prise_sum: '2000',
+    slots: '4',
+    server: 'EU',
+    start_date: '01/01/2021',
+    start_time: '8:30 p.m.',
+    info: 'bla bla bla',
+    winners: [],
+  },
+  {
+    name: 'Name Tour2',
+    id: uuidv4(),
+    url_for_banner:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606327396/useTeam/tournament2_rntabh.png',
+    url_for_card:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606327396/useTeam/tournament2_rntabh.png',
+    description: 'Its your way to become Dota 2 pro-player',
+    status: 'current',
+    mode: '1x1',
+    entry_from: '20',
+    entry_pool: 'Energy',
+    prise_pool: 'Energy',
+    prise_sum: '2000',
+    slots: '32',
+    server: 'EU',
+    start_date: '12/12/2020',
+    start_time: '8:30 p.m.',
+    info: 'bla bla bla',
+    winners: [],
+  },
+  {
+    name: 'Name Tour3',
+    id: uuidv4(),
+    url_for_banner:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606398096/useTeam/banner-dota2_z64fyi.png',
+    url_for_card:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606398096/useTeam/banner-dota2_z64fyi.png',
+    description: 'Its your way to become Dota 2 pro-player',
+    status: 'current',
+    mode: '5x5',
+    entry_from: '20',
+    entry_pool: 'Energy',
+    prise_pool: 'Energy',
+    prise_sum: '2000',
+    slots: '16',
+    server: 'USA',
+    start_date: '12/12/2020',
+    start_time: '8:30 p.m.',
+    info: 'bla bla bla',
+    winners: [],
+  },
+  {
+    name: 'Name Tour4',
+    id: uuidv4(),
+    url_for_banner:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606398101/useTeam/kunkkabloodseeker_s1bzdh.jpg',
+    url_for_card:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606398101/useTeam/kunkkabloodseeker_s1bzdh.jpg',
+    description: 'Its your way to become Dota 2 pro-player',
+    status: 'current',
+    mode: '1x1',
+    entry_from: '200',
+    entry_pool: 'Real money',
+    prise_pool: 'Real money',
+    prise_sum: '5000',
+    slots: '32',
+    server: 'EU',
+    start_date: '15/02/2021',
+    start_time: '8:30 p.m.',
+    info: 'bla bla bla',
+    winners: [],
+  },
+  {
+    name: 'Mars Arena of Universe #5',
+    id: uuidv4(),
+    url_for_banner:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606398275/useTeam/FYz_jIXNVZWCd6fDqDIRnmotJ5WCoZ-opw1FDdnxrEY_po6tao.png',
+    url_for_card:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606418730/useTeam/FYz_jIXNVZWCd6fDqDIRnmotJ5WCoZ-opw1FDdnxrEY_po6tao.png',
+    description: 'Its your way to become Dota 2 pro-player',
+    status: 'current',
+    mode: '5x5',
+    entry_from: '20',
+    entry_pool: 'Energy',
+    prise_pool: 'Energy',
+    prise_sum: '1000',
+    slots: '16',
+    server: 'EU',
+    start_date: '11/12/2020',
+    start_time: '8:30 p.m.',
+    info: 'bla bla bla',
+    winners: [],
+  },
+  {
+    name: 'Mars Arena of Universe #6',
+    id: uuidv4(),
+    url_for_banner:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606398096/useTeam/dota-2-banner-2_dweulm.jpg',
+    url_for_card:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606398096/useTeam/dota-2-banner-2_dweulm.jpg',
+    description: 'Its your way to become Dota 2 pro-player',
+    status: 'announce',
+    mode: '1x1',
+    entry_from: '20',
+    entry_pool: 'Energy',
+    prise_pool: 'Energy',
+    prise_sum: '2000',
+    slots: '32',
+    server: 'EU',
+    start_date: '12/12/2020',
+    start_time: '8:30 p.m.',
+    info: 'bla bla bla',
+    winners: [],
+  },
+  {
+    name: 'Name Tour7',
+    id: uuidv4(),
+    url_for_banner:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606398101/useTeam/kunkkabloodseeker_s1bzdh.jpg',
+    url_for_card:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606398101/useTeam/kunkkabloodseeker_s1bzdh.jpg',
+    description: 'Its your way to become Dota 2 pro-player',
+    status: 'current',
+    mode: '1x1',
+    entry_from: '200',
+    entry_pool: 'Real money',
+    prise_pool: 'Real money',
+    prise_sum: '5000',
+    slots: '32',
+    server: 'EU',
+    start_date: '15/02/2021',
+    start_time: '8:30 p.m.',
+    info: 'bla bla bla',
+    winners: [],
+  },
+  {
+    name: 'Name Tour8',
+    id: uuidv4(),
+    url_for_banner:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606398099/useTeam/dota2-banner_oug7re.jpg',
+    url_for_card:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606398099/useTeam/dota2-banner_oug7re.jpg',
+    description: 'Its your way to become Dota 2 pro-player',
+    status: 'finished',
+    mode: '1x1',
+    entry_from: '20',
+    entry_pool: 'Energy',
+    prise_pool: 'Energy',
+    prise_sum: '2000',
+    slots: '32',
+    server: 'EU',
+    start_date: '12/12/2020',
+    start_time: '8:30 p.m.',
+    info: 'bla bla bla',
+    winners: [
+      {
+        name: 'fighter',
+        rating: '234',
+      },
+    ],
+  },
+  {
+    name: 'Name Tour Big Fight 9',
+    id: uuidv4(),
+    url_for_banner:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1609686795/useTeam/Dota_2_cover_k4bdsr.jpg',
+    url_for_card:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1609686795/useTeam/Dota_2_cover_k4bdsr.jpg',
+    description: 'Its your way to become Dota 2 pro-player',
+    status: 'current',
+    mode: '1x1',
+    entry_from: '20',
+    entry_pool: 'Energy',
+    prise_pool: 'Energy',
+    prise_sum: '2000',
+    slots: '32',
+    server: 'EU',
+    start_date: '12/12/2020',
+    start_time: '8:30 p.m.',
+    info: 'bla bla bla',
+    winners: [],
+  },
+  {
+    name: 'Name Tour10',
+    id: uuidv4(),
+    url_for_banner:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1609686795/useTeam/7_image_n9sprw.jpg',
+    url_for_card:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1609686795/useTeam/7_image_n9sprw.jpg',
+    description: 'Its your way to become Dota 2 pro-player',
+    status: 'current',
+    mode: '1x1',
+    entry_from: '200',
+    entry_pool: 'Real money',
+    prise_pool: 'Real money',
+    prise_sum: '5000',
+    slots: '32',
+    server: 'EU',
+    start_date: '15/02/2021',
+    start_time: '8:30 p.m.',
+    info: 'bla bla bla',
+    winners: [],
+  },
+  {
+    name: 'Universe #11',
+    id: uuidv4(),
+    url_for_banner:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1609686795/useTeam/Dota-2-by-Valve-1280x384_ctgtsm.jpg',
+    url_for_card:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1609686795/useTeam/Dota-2-by-Valve-1280x384_ctgtsm.jpg',
+    description: 'Its your way to become Dota 2 pro-player',
+    status: 'announce',
+    mode: '1x1',
+    entry_from: '20',
+    entry_pool: 'Energy',
+    prise_pool: 'Energy',
+    prise_sum: '2000',
+    slots: '32',
+    server: 'EU',
+    start_date: '12/12/2020',
+    start_time: '8:30 p.m.',
+    info: 'bla bla bla',
+    winners: [],
+  },
+  {
+    name: 'Name Tour12',
+    id: uuidv4(),
+    url_for_banner:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1609686795/useTeam/images_lpfhmn.jpg',
+    url_for_card:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1609686795/useTeam/images_lpfhmn.jpg',
+    description: 'Its your way to become Dota 2 pro-player',
+    status: 'current',
+    mode: '1x1',
+    entry_from: '200',
+    entry_pool: 'Real money',
+    prise_pool: 'Real money',
+    prise_sum: '5000',
+    slots: '32',
+    server: 'EU',
+    start_date: '15/02/2021',
+    start_time: '8:30 p.m.',
+    info: 'bla bla bla',
+    winners: [],
+  },
+  {
+    name: 'Name Tour13',
+    id: uuidv4(),
+    url_for_banner:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1609686795/useTeam/1331783652034_2_se0fh3.webp',
+    url_for_card:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1609686795/useTeam/1331783652034_2_se0fh3.webp',
+    description: 'Its your way to become Dota 2 pro-player',
+    status: 'current',
+    mode: '1x1',
+    entry_from: '20',
+    entry_pool: 'Energy',
+    prise_pool: 'Energy',
+    prise_sum: '2000',
+    slots: '32',
+    server: 'EU',
+    start_date: '12/12/2020',
+    start_time: '8:30 p.m.',
+    info: 'bla bla bla',
+    winners: [],
+  },
+  {
+    name: 'Name Tour14',
+    id: uuidv4(),
+    url_for_banner:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606398101/useTeam/kunkkabloodseeker_s1bzdh.jpg',
+    url_for_card:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1606398101/useTeam/kunkkabloodseeker_s1bzdh.jpg',
+    description: 'Its your way to become Dota 2 pro-player',
+    status: 'current',
+    mode: '1x1',
+    entry_from: '200',
+    entry_pool: 'Real money',
+    prise_pool: 'Real money',
+    prise_sum: '5000',
+    slots: '32',
+    server: 'EU',
+    start_date: '15/02/2021',
+    start_time: '8:30 p.m.',
+    info: 'bla bla bla',
+    winners: [],
+  },
+  {
+    name: 'Name Tour15',
+    id: uuidv4(),
+    url_for_banner:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1609686795/useTeam/dota-2-alchemist-facebook-cover-timeline-banner-for-fb_zmuacx.jpg',
+    url_for_card:
+      'https://res.cloudinary.com/dk88eyahu/image/upload/v1609686795/useTeam/dota-2-alchemist-facebook-cover-timeline-banner-for-fb_zmuacx.jpg',
+    description: 'Its your way to become Dota 2 pro-player',
+    status: 'current',
+    mode: '1x1',
+    entry_from: '20',
+    entry_pool: 'Energy',
+    prise_pool: 'Energy',
+    prise_sum: '2000',
+    slots: '32',
+    server: 'EU',
+    start_date: '12/12/2020',
+    start_time: '8:30 p.m.',
+    info: 'bla bla bla',
+    winners: [],
+  },
+];
+
+export const InfoListData = [
+  {
+    id: uuidv4(),
+    name: 'PAY LESS FOR INVITE',
+    description: 'Team save 20% on each tournament invite',
+  },
+  {
+    id: uuidv4(),
+    name: 'GET MORE PRIZE',
+    description: 'Team owner get 30% instead 20% when play solo',
+  },
+  {
+    id: uuidv4(),
+    name: 'FIND NEW FRIENDS',
+    description: 'Every victory is a team work',
+  },
+];
+
+export const GetTour = (status, numElem) =>
+  TournamentsData.filter((el, index) =>
+    numElem ? index <= numElem && el.status === status : el.status === status,
+  );
