+++ src/client/mainPage/pages/MainPage/style.jsx
