+++ src/shared/components/Loaders/Loaders.module.scss
@@ -1,0 +1,12 @@
+.buttonContainer {
+    display: flex;
+    justify-content: center;
+    margin-top: 10px;
+}
+
+.loadMoreHide{
+    display: none;
+}
+.loadMoreVisible{
+    visibility: visible;
+}
