+++ src/shared/components/Loaders/index.js
@@ -1,0 +1,1 @@
+export {default} from './Loaders'
